+++ src/indexers/hosted/indexer.ts
